Agregando excepciones + Actualizando tests + DTO Turno

+++ src/main/java/com/dh/TPIdiego_cata/service/implementation/OdontologoService.java
@@ -1,8 +1,10 @@
 package com.dh.TPIdiego_cata.service.implementation;
 
 import com.dh.TPIdiego_cata.entity.Odontologo;
+import com.dh.TPIdiego_cata.exceptions.ResourceNotFoundException;
 import com.dh.TPIdiego_cata.repository.IOdontologoRepository;
 import com.dh.TPIdiego_cata.service.IOdontologoService;
+import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -12,6 +14,8 @@ import java.util.Optional;
 @Service
 public class OdontologoService implements IOdontologoService {
 
+    private static final Logger logger = Logger.getLogger(OdontologoService.class);
+
     private IOdontologoRepository odontologoRepository;
     @Autowired
     public OdontologoService(IOdontologoRepository odontologoRepository) {
@@ -20,31 +24,54 @@ public class OdontologoService implements IOdontologoService {
 
     @Override
     public Odontologo guardar(Odontologo odontologo) {
+        logger.info("Guardando un odontólogo");
         return odontologoRepository.save(odontologo);
     }
 
     @Override
     public Odontologo buscarPorId(Long id) {
-        Odontologo odontologo = null;
+        logger.info("Buscando odontólogo con ID "+id);
         Optional<Odontologo> odontologoOptional = odontologoRepository.findById(id);
         if(odontologoOptional.isPresent()) {
-            odontologo=odontologoOptional.get();
+            return odontologoOptional.get();
+        }else{
+            String message = "No se ha encontrado el odontologo con ID "+id;
+            logger.error(message);
+            throw new ResourceNotFoundException(message);
         }
-        return odontologo;
     }
 
     @Override
     public void actualizar(Odontologo odontologo) {
-        odontologoRepository.save(odontologo);
+        logger.info("Actualizando odontólogo");
+        Optional<Odontologo> odontologoBuscado = odontologoRepository.findById(odontologo.getId());
+        if (odontologoBuscado.isPresent()) {
+            odontologoRepository.save(odontologo);
+            logger.info("Se guardó odontólogo actualizado");
+        }else{
+            String message = "No se ha encontrado el odontologo a actualizar con ID "+odontologo.getId();
+            logger.error(message);
+            throw new ResourceNotFoundException(message);
+        }
     }
 
     @Override
     public void eliminar(Long id) {
-        odontologoRepository.deleteById(id);
+        logger.info("Eliminando odontólogo");
+        Optional<Odontologo> odontologoAEliminar = odontologoRepository.findById(id);
+        if (odontologoAEliminar.isPresent()) {
+            odontologoRepository.deleteById(id);
+            logger.info("Se eliminó el odontólgo con ID"+id);
+        }else{
+            String message = "No se ha encontrado el odontologo a eliminar con ID "+id;
+            logger.error(message);
+            throw new ResourceNotFoundException(message);
+        }
     }
 
     @Override
     public List<Odontologo> listarTodos() {
+        logger.info("Listando todos los odontólogo");
         return odontologoRepository.findAll();
     }
 }
